Split Descriptor in sub module
